web-app/requirements.txt
aniso8601==4.0.1
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
enum34==1.1.6
Flask==1.0.2
Flask-JWT-Extended==3.13.1
flask-restplus==0.12.1
functools32==3.2.3.post2
idna==2.7
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
py-bcrypt==0.4
pycparser==2.19
pycryptodome==3.7.0
PyJWT==1.6.4
python-dateutil==2.7.5
python-redis-lock==3.2.0
pytz==2018.7
redis==2.10.6
requests==2.20.0
rom==0.42.3
six==1.11.0
urllib3==1.24
virtualenv==16.0.0
Werkzeug==0.14.1
